+++ src/appearance/theme/selectable_text.rs
@@ -138,18 +138,25 @@ pub fn nickname(
     is_away: bool,
     is_user_offline: bool,
 ) -> Style {
-    let is_away = config
-        .buffer
-        .nickname
-        .away
-        .is_away(is_away || is_user_offline);
+    let is_offline = config.buffer.nickname.offline.style(is_user_offline);
+    let is_away = config.buffer.nickname.away.is_away(is_away);
 
     if let Some(metadata_color) = metadata_color {
-        let color = nickname_alpha(
-            metadata_color,
-            is_away,
-            theme.styles().buffer.background,
-        );
+        let background = theme.styles().buffer.background;
+        let color = if let Some(style) = is_offline {
+            let color = match style.color {
+                buffer::nickname::OfflineColor::Theme => theme
+                    .styles()
+                    .buffer
+                    .nickname_offline
+                    .color
+                    .unwrap_or(metadata_color),
+                buffer::nickname::OfflineColor::Nickname => metadata_color,
+            };
+            nickname_alpha(color, Some(style.alpha), background)
+        } else {
+            nickname_alpha(metadata_color, is_away, background)
+        };
 
         Style {
             color: Some(color),
@@ -161,7 +168,7 @@ pub fn nickname(
             &config.buffer.nickname.color,
             user,
             is_away,
-            config.buffer.nickname.offline.is_offline(is_user_offline),
+            is_offline,
         )
     }
 }
@@ -170,8 +177,8 @@ fn nickname_style(
     theme: &Theme,
     kind: &data::buffer::Color,
     user: &User,
-    is_away: Option<buffer::Away>,
-    is_offline: bool,
+    is_away: Option<buffer::Alpha>,
+    is_offline: Option<buffer::nickname::OfflineStyle>,
 ) -> Style {
     let color =
         text::nickname(theme, kind, Some(user.seed()), is_away, is_offline)